Backend/requirements.txt
tensorflow==2.10.1; sys_platform == 'win32' and python_version < '3.11'
tensorflow>=2.15.0; sys_platform != 'win32' or python_version >= '3.11'
numpy>=1.26.0,<2.0.0
Pillow>=10.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
pydantic>=2.0.0
# Ensure Ollama is installed and running locally
langchain-ollama>=0.0.1
firebase-admin>=6.5.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-talisman>=1.1.0
requests>=2.31.0
streamlit>=1.30.0
protobuf==3.19.6; sys_platform == 'win32' and python_version < '3.11'
protobuf==3.20.3; sys_platform != 'win32' or python_version >= '3.11'
psutil>=5.9.0
httpx>=0.23.0
